+++ .azure-devops/templates/nightly-tests.yml
@@ -12,86 +12,83 @@ parameters:
   - min
   - released
   - edge
-- name: num_threads
-  type: number
-  default: 6
-- name: num_reruns
-  type: number
-  default: 2
-- name: reruns_delay
-  type: number
-  default: 60
+- name: azureSubscription
+  type: string
+  default: 'az_cli_nightly_conn'
 
-steps:
-  - template: setup-dev-test-env.yml
-    parameters:
-      architecture: ${{ parameters.architecture }}
-      pythonVersion: ${{ parameters.pythonVersion }}
-      azureCLIVersion: ${{ parameters.azureCLIVersion }}
+stages:
+  - stage: 'setup'
+    displayName: 'Setup dev test environment'
+    jobs:
 
-  - template: set-testenv-sentinel.yml
+    - job: 'Setup_Test_Environment'
+      pool:
+        vmImage: 'Ubuntu-16.04'
 
-  - script: |
-      pytest -vv azext_iot/tests -k "_unit" --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-unit.xml
-    displayName: 'All unit tests'
-    env:
-      COVERAGE_FILE: .coverage.all
+      steps:
+      - template: setup-dev-test-env.yml
+        parameters:
+          architecture: ${{ parameters.architecture }}
+          pythonVersion: ${{ parameters.pythonVersion }}
+          azureCLIVersion: ${{ parameters.azureCLIVersion }}
 
-  - task: AzureCLI@2
-    continueOnError: true
-    displayName: 'IoTHub integration tests - devices, messaging and modules'
-    inputs:
-      azureSubscription: az_cli_nightly_conn
-      scriptType: bash
-      scriptLocation: inlineScript
-      inlineScript: |
-        export COVERAGE_FILE=.coverage.all
-        pytest -vv azext_iot/tests/iothub/devices azext_iot/tests/iothub/messaging azext_iot/tests/iothub/modules -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml
+      - template: set-testenv-sentinel.yml
 
-  - task: AzureCLI@2
-    continueOnError: true
-    displayName: 'IoTHub integration tests - jobs, hub and configuration'
-    inputs:
-      azureSubscription: az_cli_nightly_conn
-      scriptType: bash
-      scriptLocation: inlineScript
-      inlineScript: |
-        export COVERAGE_FILE=.coverage.all
-        pytest -vv azext_iot/tests/iothub --ignore=azext_iot/tests/iothub/devices --ignore=azext_iot/tests/iothub/messaging --ignore=azext_iot/tests/iothub/modules -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml --deselect=azext_iot/tests/iothub/test_iothub_discovery_int.py::TestIoTHubDiscovery::test_iothub_targets
-        
-  - task: AzureCLI@2
-    continueOnError: true
-    displayName: 'Central, DPS and Product integration tests'
-    inputs:
-      azureSubscription: az_cli_nightly_conn
-      scriptType: bash
-      scriptLocation: inlineScript
-      inlineScript: |
-        export COVERAGE_FILE=.coverage.all
-        pytest -vv azext_iot/tests/central azext_iot/tests/dps azext_iot/tests/product -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml
+  - stage: 'test'
+    displayName: 'Run tests'
+    pool:
+      vmImage: 'Ubuntu-16.04'
+    dependsOn: build
+    jobs:
+    - job: 'testCentral'
+      displayName: 'Test IoT Central'
+      steps:
+      - template: templates/run-tests-parallel.yml
+        parameters:
+          path: 'azext_iot/tests/central'
+          name: 'iot-central'
+          azureSubscription: ${{ parameters.azureSubscription }}
+    
+    - job: 'testADT'
+      displayName: 'Test Azure DigitalTwins'
+      steps:
+      - template: templates/run-tests-parallel.yml
+        parameters:
+          path: 'azext_iot/tests/digitaltwins'
+          name: 'azure-digitaltwins'
+          azureSubscription: ${{ parameters.azureSubscription }}
 
-  - task: AzureCLI@2
-    continueOnError: true
-    displayName: 'Digital Twins integration tests'
-    inputs:
-      azureSubscription: az_cli_nightly_conn
-      scriptType: bash
-      scriptLocation: inlineScript
-      inlineScript: |
-        export COVERAGE_FILE=.coverage.all
-        pytest -vv azext_iot/tests/digitaltwins -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml
-        
-  - task: PublishBuildArtifacts@1
-    inputs:
-      pathToPublish: .coverage.all
-      publishLocation: 'Container' 
-      artifactName: 'coverage'
+    - job: 'testDPS'
+      displayName: 'Test DPS'
+      steps:
+      - template: templates/run-tests-parallel.yml
+        parameters:
+          path: 'azext_iot/tests/dps'
+          name: 'device-provisioning-service'
+          azureSubscription: ${{ parameters.azureSubscription }}
+    
+    - job: 'testHub_job_1'
+      displayName: 'Test IoT Hub - config, core and jobs'
+      steps:
+      - template: templates/run-tests-parallel.yml
+        parameters:
+          path: 'azext_iot/tests/iothub/configurations azext_iot/tests/iothub/core azext_iot/tests/iothub/jobs'
+          name: 'iot-hub-1'
+          azureSubscription: ${{ parameters.azureSubscription }}
+    
+    - job: 'testHub_job_2'
+      displayName: 'Test IoT Hub - devices, messaging and modules'
+      steps:
+      - template: templates/run-tests-parallel.yml
+        parameters:
+          path: 'azext_iot/tests/iothub/devices azext_iot/tests/iothub/messaging azext_iot/tests/iothub/modules'
+          name: 'iot-hub-2'
+          azureSubscription: ${{ parameters.azureSubscription }}
 
-  - task: PublishTestResults@2
-    condition: succeededOrFailed()
-    displayName: 'Publish Test Results'
-    inputs:
-      testResultsFormat: 'JUnit'
-      testResultsFiles: '**/test-*.xml'
-      testRunTitle: 'Publish test results for Python ${{ parameters.pythonVersion }} on OS $(Agent.OS)'
-      searchFolder: '$(System.DefaultWorkingDirectory)'
+    - job: 'unitTests'
+      displayName: 'Unit tests and code coverage'
+      steps:
+      - template: templates/run-tests-parallel.yml
+        parameters:
+          runIntTests: 'false'
+          runUnitTests: 'true'
